+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 
 # Change Log
 
+## 2018-12-03
+* 新增 ccIns.call, ccIns.callWith函数，方便更灵活的改变state
+* 修复在webpack 热加载模式下，CC_CONTEXT.refs 维护不正确的问题